refactor filter handling to a hook
